Implement pattern data calculation

+++ src/main/java/com/group/atelier/model/entity/ClientMetrics.java
@@ -18,50 +18,50 @@ public class ClientMetrics {
     private int clientId;
 
     @Column(name = "neck_semi_circumference")
-    private float neckSemiCircumference;
+    private Double neckSemiCircumference;
 
     @Column(name = "chest_semi_circumference_1")
-    private float chestSemiCircumference1;
+    private Double chestSemiCircumference1;
 
     @Column(name = "chest_semi_circumference_2")
-    private float chestSemiCircumference2;
+    private Double chestSemiCircumference2;
 
     @Column(name = "chest_semi_circumference_3")
-    private float chestSemiCircumference3;
+    private Double chestSemiCircumference3;
 
     @Column(name = "waist_semi_circumference")
-    private float waistSemiCircumference;
+    private Double waistSemiCircumference;
 
     @Column(name = "shoulder_width")
-    private float shoulderWidth;
+    private Double shoulderWidth;
 
     @Column(name = "chest_height")
-    private float chestHeight;
+    private Double chestHeight;
 
     @Column(name = "chest_height_1")
-    private float chestHeight1;
+    private Double chestHeight1;
 
     @Column(name = "back_armhole_height")
-    private float backArmholeHeight;
+    private Double backArmholeHeight;
 
     @Column(name = "back_length_till_waist")
-    private float backLengthTillWaist;
+    private Double backLengthTillWaist;
 
     @Column(name = "shoulder_height_sidelong")
-    private float shoulderHeightSidelong;
+    private Double shoulderHeightSidelong;
 
     @Column(name = "chest_width")
-    private float chestWidth;
+    private Double chestWidth;
 
     @Column(name = "chest_center")
-    private float chestCenter;
+    private Double chestCenter;
 
     @Column(name = "back_width")
-    private float backWidth;
+    private Double backWidth;
 
     @Column(name = "waist_length_front")
-    private float waistLengthFront;
+    private Double waistLengthFront;
 
     @Column(name = "neck_base_to_front_waist_line_distance")
-    private float neckBaseToFrontWaistLineDistance;
+    private Double neckBaseToFrontWaistLineDistance;
 }
